Small fixes
git-svn-id: file:///Users/tack/GecodeGitMigration/gecode-svn-mirror/gecode/trunk@4897 e85b7adc-8362-4630-8c63-7469d557c915

+++ gecode/support/dynamic-array.hh
@@ -84,7 +84,7 @@ namespace Gecode { namespace Support {
   template <class T>
   forceinline
   DynamicArray<T>::DynamicArray(int m)
-    : n(m), x(Memory::bmalloc<T>(n)) {}
+    : n(m), x(Memory::bmalloc<T>(static_cast<size_t>(n))) {}
 
   template <class T>
   forceinline
